requirements.txt
asgiref==3.7.2
Django~=3.2.12
form==0.0.1
sqlparse~=0.4.2
typing_extensions==4.9.0

jupyter~=1.0.0
pip~=22.0.2
attrs~=23.2.0
wheel~=0.37.1
cryptography~=41.0.7
pillow~=10.2.0
tornado~=6.4
Jinja2~=3.1.3
uuid~=1.30
keyring~=23.5.0
colorama~=0.4.4
PyYAML~=5.4.1
numpy~=1.26.3
requests~=2.31.0
setuptools~=59.6.0
future~=0.18.2
PyJWT~=2.3.0
djangorestframework~=3.14.0